Removing pid_handler.

+++ lib/vagrant-notify-forwarder/action/stop_host_forwarder.rb
@@ -1,5 +1,4 @@
 require 'vagrant-notify-forwarder/utils'
-require 'vagrant-notify-forwarder/pid_handler'
 
 module VagrantPlugins
   module VagrantNotifyForwarder
@@ -7,23 +6,27 @@ module VagrantPlugins
       class StopHostForwarder
         def initialize(app, env)
           @app = app
-          @pid_handler = PidHandler.instance
         end
 
         def call(env)
           @app.call env
 
           return unless env[:machine].config.notify_forwarder.enable
-          env[:ui].info "Processes to stop are #{@pid_handler.pids}"
-          @pid_handler.pids.each { |pid|
-            begin
-              env[:ui].info "Stopping process #{pid}"
-              Process.kill 'TERM', pid
-            rescue Errno::ESRCH
-            end
-          }
 
-          @pid_handler.pids = []
+          pidfile = Utils.host_pidfile env
+
+          if File.exists? pidfile
+            pidfile.open('r') do |f|
+              f.readlines.each do |process|
+                pid = process.to_i
+                begin
+                  Process.kill 'TERM', pid
+                rescue Errno::ESRCH
+                end
+              end
+              pidfile.delete
+            end
+          end
         end
       end
     end
